move TeXFont to a individual file
