+++ src/fonts/fonts.h
@@ -11,233 +11,13 @@
 #include "core/formula.h"
 #include "fonts/alphabet.h"
 #include "fonts/font_info.h"
+#include "fonts/tex_font.h"
 #include "graphic/graphic.h"
 
 using namespace std;
 using namespace tex;
 
 namespace tex {
-
-/**
- * An interface representing a "TeXFont", which is responsible for all the
- * necessary fonts and font information.
- */
-class TeXFont {
-public:
-  static const int NO_FONT;
-
-  virtual float getScaleFactor() = 0;
-
-  virtual float getAxisHeight(int style) = 0;
-
-  virtual float getBigOpSpacing1(int style) = 0;
-
-  virtual float getBigOpSpacing2(int style) = 0;
-
-  virtual float getBigOpSpacing3(int style) = 0;
-
-  virtual float getBigOpSpacing4(int style) = 0;
-
-  virtual float getBigOpSpacing5(int style) = 0;
-
-  /**
-   * Get a Char-object specifying the given character in the given text style
-   * with metric information depending on the given "style".
-   * @param c
-   *      alphanumeric character
-   * @param textStyle
-   *      the text style in which the character should be drawn
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return
-   *      the Char-object specifying the given character in the given text style
-   * @throw ex_text_style_mapping_not_found
-   *      if the text style mapping not found
-   */
-  virtual Char getChar(
-      wchar_t       c,
-      const string& textStyle,
-      int           style) throw(ex_text_style_mapping_not_found) = 0;
-
-  /**
-   * Get a Char-object for this specific character containing the metric information
-   * @param cf
-   *      CharFont-object determine a specific character of a specific font
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return the Char-object for this character containing metric information
-   */
-  virtual Char getChar(const CharFont& cf, int style) = 0;
-
-  /**
-   * Get a Char-object for the given symbol with metric information depending on "style"
-   * @param name
-   *      the symbol name
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return a Char-object for this symbol with metric information
-   */
-  virtual Char getChar(const string& name, int style) throw(ex_symbol_mapping_not_found) = 0;
-
-  /**
-   * Get a Char-object specifying the given character in the default text style
-   * with metric information depending on the given "style"
-   * @param c
-   *      alphanumeric character
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return the Char-object specifying the given character in the default text style
-   */
-  virtual Char getDefaultChar(wchar_t c, int style) = 0;
-
-  virtual float getDefaultRuleThickness(int style) = 0;
-
-  virtual float getDenom1(int style) = 0;
-
-  virtual float getDenom2(int style) = 0;
-
-  /**
-   * Get an Extension-object for the given Char containing the 4 possible
-   * parts to build an arbitrary large variant. This will only be called if
-   * isExtensionChar(Char) returns true.
-   * @param c
-   *      a Char-object for a specific character
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return an extension object containing the 4 possible parts
-   */
-  virtual Extension* getExtension(const Char& c, int style) = 0;
-
-  /**
-   * Get the kern value to be inserted between the given characters in the
-   * given style
-   * @param left
-   *      left character
-   * @param right
-   *      right character
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return the kern value between both characters (default 0)
-   */
-  virtual float getKern(const CharFont& left, const CharFont& right, int style) = 0;
-
-  /**
-   * Get the ligature that replaces both characters (if any).
-   * @param left
-   *      left character
-   * @param right
-   *      right character
-   * @return a ligature replacing both characters (or null if no any ligature)
-   */
-  virtual sptr<CharFont> getLigature(const CharFont& left, const CharFont& right) = 0;
-
-  /**
-   * Get the id of mu font
-   */
-  virtual int getMuFontId() = 0;
-
-  /**
-   * Get the next larger version of the given character. This is only called
-   * if hasNextLarger(Char) returns true.
-   * @param c
-   *      specified character
-   * @param style
-   *      the style in which the atom should be drawn
-   * @return the next larger version of this character
-   */
-  virtual Char getNextLarger(const Char& c, int style) = 0;
-
-  virtual float getNum1(int style) = 0;
-
-  virtual float getNum2(int style) = 0;
-
-  virtual float getNum3(int style) = 0;
-
-  virtual float getQuad(int style, int fontCode) = 0;
-
-  /**
-   * @return the point size of this TeXFont
-   */
-  virtual float getSize() = 0;
-
-  /**
-   * Get the kern amount of the character defined by the given CharFont
-   * followed by the "skewChar" of it's font. This is used in the algorithm
-   * for placing an accent above a single character.
-   * @param cf
-   *      the character an it's font above which an accent has to be placed
-   * @param style
-   *      the render style
-   * @return the kern amount of the character defined by cf followed by the
-   * "skewChar" of it's font
-   */
-  virtual float getSkew(const CharFont& cf, int style) = 0;
-
-  virtual float getSpace(int style) = 0;
-
-  virtual float getSub1(int style) = 0;
-
-  virtual float getSub2(int style) = 0;
-
-  virtual float getSubDrop(int style) = 0;
-
-  virtual float getSup1(int style) = 0;
-
-  virtual float getSup2(int style) = 0;
-
-  virtual float getSup3(int style) = 0;
-
-  virtual float getSupDrop(int style) = 0;
-
-  virtual float getXHeight(int style, int fontCode) = 0;
-
-  virtual float getEM(int style) = 0;
-
-  /**
-   * Test if a character has larger version.
-   * 
-   * @param c
-   *      a character
-   * @return true if the given character has a larger version, false otherwise
-   */
-  virtual bool hasNextLarger(const Char& c) = 0;
-
-  virtual bool hasSpace(int font) = 0;
-
-  virtual void setBold(bool bold) = 0;
-
-  virtual bool getBold() = 0;
-
-  virtual void setRoman(bool rm) = 0;
-
-  virtual bool getRoman() = 0;
-
-  virtual void setTt(bool tt) = 0;
-
-  virtual bool getTt() = 0;
-
-  virtual void setIt(bool it) = 0;
-
-  virtual bool getIt() = 0;
-
-  virtual void setSs(bool ss) = 0;
-
-  virtual bool getSs() = 0;
-
-  /**
-   * Test if the given character is an extension character.
-   * 
-   * @param c
-   *      a character
-   * @return true if the given character contains extension information to
-   * build an arbitrary large version of this character.
-   */
-  virtual bool isExtensionChar(const Char& c) = 0;
-
-  virtual sptr<TeXFont> copy() = 0;
-
-  virtual ~TeXFont();
-};
 
 typedef struct {
   int    font;
@@ -438,23 +218,23 @@ public:
 
   inline void setBold(bool bold) override { _isBold = bold; }
 
-  inline bool getBold() override { return _isBold; }
+  inline bool isBold() override { return _isBold; }
 
   inline void setRoman(bool rm) override { _isRoman = rm; }
 
-  inline bool getRoman() override { return _isRoman; }
+  inline bool isRoman() override { return _isRoman; }
 
   inline void setTt(bool tt) override { _isTt = tt; }
 
-  inline bool getTt() override { return _isTt; }
+  inline bool isTt() override { return _isTt; }
 
   inline void setSs(bool ss) override { _isSs = ss; }
 
-  inline bool getSs() override { return _isSs; }
+  inline bool isSs() override { return _isSs; }
 
   inline void setIt(bool it) override { _isIt = it; }
 
-  inline bool getIt() override { return _isIt; }
+  inline bool isIt() override { return _isIt; }
 
   inline bool hasSpace(int font) override {
     FontInfo* info = getInfo(font);
